requirements.txt
torch==2.5.1
pillow==11.0.0
uvicorn==0.32.1
streamlit-1.40.2
fastapi==0.115.5
torchvision==0.20.1
python-multipart==0.0.17